+++ packages/terra-clinical-data-grid/src/subcomponents/Cell.jsx
@@ -20,7 +20,7 @@ const propTypes = {
    */
   sectionId: PropTypes.string.isRequired,
   /**
-   * String identifer of the row in which the Cell will be rendered.
+   * String identifier of the row in which the Cell will be rendered.
    */
   rowId: PropTypes.string.isRequired,
   /**
@@ -62,6 +62,18 @@ const propTypes = {
    * Function that will be called with a ref to the Cell's selectable element. Parameters: `selectableRefCallback(selectableRef)`
    */
   selectableRefCallback: PropTypes.func,
+  /**
+   * Boolean indicating whether the cell should be highlighted with background styling.
+   */
+  isColumnHighlighted: PropTypes.bool,
+  /**
+   * Boolean indicating whether the cell is in the first row, also used with column highlight background styling.
+   */
+  isFirstRow: PropTypes.bool,
+  /**
+   * Boolean indicating whether the cell is in the last row, also used with column highlight background styling.
+   */
+  isLastRow: PropTypes.bool,
 };
 
 class Cell extends React.Component {
@@ -119,6 +131,9 @@ class Cell extends React.Component {
       onHoverStart,
       onHoverEnd,
       ariaLabel,
+      isColumnHighlighted,
+      isFirstRow,
+      isLastRow,
       ...customProps
     } = this.props;
 
@@ -126,6 +141,14 @@ class Cell extends React.Component {
     const theme = this.context;
     const role = isSelectable ? 'button' : undefined;
     const tabIndex = isSelectable ? '0' : undefined;
+
+    const highlightCellClassNames = !isSelected && isColumnHighlighted && cx([
+      { highlighted: !isSelectable },
+      { 'highlighted-selectable': isSelectable },
+      { first: isFirstRow },
+      { last: isLastRow },
+    ]);
+
     return (
       <div
         {...customProps}
@@ -135,7 +158,12 @@ class Cell extends React.Component {
       >
         <div
           role={role}
-          className={cx(['content', { selectable: isSelectable, selected: isSelected }])}
+          className={cx([
+            'content',
+            highlightCellClassNames,
+            { selectable: isSelectable },
+            { selected: isSelected },
+          ])}
           onClick={isSelectable ? this.handleTargetClick : undefined}
           onKeyDown={isSelectable ? this.handleKeyDown : undefined}
           onMouseEnter={onHoverStart}