+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added new prop `columnHighlightId` to highlight a column with style intended for use in timelines
+
 ## 2.34.0 - (November 16, 2021)
 
 * Changed